Add files via upload
We use R and apply exploratory data analysis techniques to explore relationships in one variable to multiple variables and to explore a selected data set for distributions, outliers, and anomalies.

The Following is an analysis of data made on Prosper Company's Loan Prediction data set. Prosper is America's first marketplace lending platform. Get a personal loan at a low rate. Prosper personal loans require generally good credit; this peer-to-peer lender grades your loan so investors can decide whether to fund it.

Here are Some of the features of the loan offered by Prosper:

Low interest rate
Fixed term-3 or 5 years*
Single monthly payment
No hidden fees or prepayment penalties
The Prosper Loans dataset contains a wealth of information about 114K loans given between Nov 2005 and Mar 2014.

I began my investigation by going through each of the variables. I then categorized them into variables that are useful and should be taken into consideration and the ones that not be taken under consideration.

I continued exploring individual variables that I had noticed, like the loan amount, term, interest rates, credit scores, loan status, and loan purposes. At first I leaned towards the relationship between the interest rates and the credit score, but because I found it so evident I decided to give it another thought and I read again the definition of the variables. I thought backwards, like a borrower, and decided to explore what would give me a good or bad credit rating that would deserve me a lower interest rate.

Then I went on to decide the types of plots that would be apt for depicting the models in a clear and understandable way. I went through a few difficulties in plotting the models, especially when i took 3 variables into consideration. Another difficulty was that some of the variables had no corelation and it was difficult to express them in a graph and derive the results.